refactor: rename pantry add client route
